+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.1.5]
+
+### _Feature_
+
+- Wearable API: Now you can add wearable images to an image,
+  child image will be rendered relative to the parent image.
+
+### Fixed
+
+- Throw error when deserializing a game state with image state
+
+### Added
+
+- Wearable API
+
 ## [0.1.3] - 2024/10/29
 
 ### _Feature_